refactor(pubsublite): rename common.PublishMetadata to publish.Metadata
