+++ pubsublite/internal/wire/publish_batcher.go
@@ -18,7 +18,7 @@ import (
 	"errors"
 	"fmt"
 
-	"cloud.google.com/go/pubsublite/common"
+	"cloud.google.com/go/pubsublite/publish"
 	"google.golang.org/api/support/bundler"
 	"google.golang.org/protobuf/proto"
 
@@ -28,7 +28,7 @@ import (
 var errPublishQueueEmpty = errors.New("pubsublite: received publish response from server with no batches in flight")
 
 // PublishResultFunc receives the result of a publish.
-type PublishResultFunc func(*common.PublishMetadata, error)
+type PublishResultFunc func(*publish.Metadata, error)
 
 // messageHolder stores a message to be published, with associated metadata.
 type messageHolder struct {
@@ -141,7 +141,7 @@ func (b *publishMessageBatcher) OnPublishResponse(firstOffset int64) error {
 	batch, _ := frontElem.Value.(*publishBatch)
 	for i, msgHolder := range batch.msgHolders {
 		// Messages are ordered, so the offset of each message is firstOffset + i.
-		pm := &common.PublishMetadata{Partition: b.partition, Offset: firstOffset + int64(i)}
+		pm := &publish.Metadata{Partition: b.partition, Offset: firstOffset + int64(i)}
 		msgHolder.onResult(pm, nil)
 		b.availableBufferBytes += msgHolder.size
 	}
