+++ pubsublite/publish/metadata_test.go
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 
-package common
+package publish
 
 import (
 	"testing"
@@ -23,18 +23,18 @@ func TestPublishMetadataStringEncoding(t *testing.T) {
 	for _, tc := range []struct {
 		desc    string
 		input   string
-		want    *PublishMetadata
+		want    *Metadata
 		wantErr bool
 	}{
 		{
 			desc:  "valid: zero",
 			input: "0:0",
-			want:  &PublishMetadata{Partition: 0, Offset: 0},
+			want:  &Metadata{Partition: 0, Offset: 0},
 		},
 		{
 			desc:  "valid: non-zero",
 			input: "3:1234",
-			want:  &PublishMetadata{Partition: 3, Offset: 1234},
+			want:  &Metadata{Partition: 3, Offset: 1234},
 		},
 		{
 			desc:    "invalid: number",
@@ -53,14 +53,14 @@ func TestPublishMetadataStringEncoding(t *testing.T) {
 		},
 	} {
 		t.Run(tc.desc, func(t *testing.T) {
-			got, gotErr := ParsePublishMetadata(tc.input)
+			got, gotErr := ParseMetadata(tc.input)
 			if !testutil.Equal(got, tc.want) || (gotErr != nil) != tc.wantErr {
-				t.Errorf("ParsePublishMetadata(%q): got (%v, %v), want (%v, err=%v)", tc.input, got, gotErr, tc.want, tc.wantErr)
+				t.Errorf("ParseMetadata(%q): got (%v, %v), want (%v, err=%v)", tc.input, got, gotErr, tc.want, tc.wantErr)
 			}
 
 			if tc.want != nil {
 				if got := tc.want.String(); got != tc.input {
-					t.Errorf("PublishMetadata(%v).String(): got %q, want: %q", tc.want, got, tc.input)
+					t.Errorf("Metadata(%v).String(): got %q, want: %q", tc.want, got, tc.input)
 				}
 			}
 		})
